requirements.txt
Django==1.3.7
MySQL-python==1.2.5
Pillow==2.3.0
argparse==1.2.1
distribute==0.6.24
django-extensions==0.8
django-mptt==0.5.2
django-robots==1.0
django-simple-captcha==0.3.0
django-tagging==0.3.1
gunicorn==18.0
wsgiref==0.1.2
